requirements.txt
# server_py/requirements.txt
fastapi
uvicorn[standard]
python-dotenv
nba_api
pytz
# Podríamos necesitar 'requests' si hacemos llamadas directas, pero nba_api debería bastar